Group summary rows & autosaved column properties
- Replaced the TFUserDefaults dependency with
-tableGridAutosavedColumnProperties: and
-tableGrid:didAutosaveColumnProperties: delegate methods.
- Added -tableGrid:groupSummaryCellForColumn:row: data source method,
to provide the group summary cell (similar to the other equivalent
methods),
- Added -tableGrid:updateGroupSummaryCell:forColumn:row: data source
method, to enable changing the attributes of the cell (e.g. font and
text color).
- Added -tableGrid:groupSummaryValueForColumn:row: data source method,
to provide the value for the group summary cell.
- Added -tableGrid:groupSummaryBackgroundColorForColumn:row: data
source method, to enable changing the background color.
- The table grid now uses those and the -tableGrid:isGroupRow: one
(which still indicates the group heading rows) to populate and draw the
group summary rows.  Note that you should allow for the additional
summary rows when populating the data source.
- Added an includeGroupSummaryRows boolean property to MBTableGrid.
Defaults to YES.  If set to NO, the group summary rows are omitted.
Note that after toggling you need to adjust the data source (to account
for the summary rows) and reload the table grid.
- See the demo project for example usage, including a new Table menu
item to toggle the summary rows, and support for negatives in red.